+++ firmware/copter/source/communicator/GlobalPositionInfo.hpp
@@ -11,14 +11,14 @@ namespace copter::communicator
 struct GlobalPositionInfo
 {
     std::uint32_t time_boot_ms = {};
-    float latitude_deg = {};
-    float longitude_deg = {};
-    float altitude_m = {};
-    float relative_altitude_m = {};
-    float speed_x_ms = {};
-    float speed_y_ms = {};
-    float speed_z_ms = {};
-    float yaw_deg = {};
+    double latitude_deg = {};
+    double longitude_deg = {};
+    double altitude_m = {};
+    double relative_altitude_m = {};
+    double speed_x_ms = {};
+    double speed_y_ms = {};
+    double speed_z_ms = {};
+    double yaw_deg = {};
 
     GlobalPositionInfo(const __mavlink_global_position_int_t& global_position);
 };
